exim: remove Received headers in mails submitted by authenticated users
This should prevent privacy issues (do not disclose the user's ip address)
but is also part of the DDOS protection (if our website is protected by
something like CloudFlare the attacker must not be able to get our real
IP - but that IP would appear in email headers e.g. for new user
registrations).