common: ability to use a custom name for the eth device (instead of hardcoding eth0)
CentOS tries to calculate a unique name for each eth device instead of using
generic names like eth0, eth1, ... So while these names are stable even when a new
card is plugged in this means we don't have the same eth names accross different
hosts.

As we need to reference the actual eth device in very few places it seems best
just to adapt to that change instead of forcing Linux to continue using the old
names.

The networkd file name in the "host_files" directory is still named "10-eth0.network"
so I don't have to rename the file for all hosts (the file will be stored with
a different name on the provisioned host).